driver-api/requirements.txt
flask==2.0.1
dapr==1.8.0
kafka-python==2.0.2
psycopg2-binary==2.9.3